Fix bugs found in codebase audit
- route_matrix.go: NoRoutePercent and OneRoutePercent used integer
  division, always reporting 0% - now use float64 division
- relay_packets.go: RelayUpdateRequestPacket version check used &&
  instead of ||, so it could never fail - out-of-range versions are
  now rejected
- encoding.go: ReadAddress now bounds-checks the address type byte
  (no panic on truncated packets), returns true for a valid None
  address instead of failing, and reads IPv6 as exactly 16 bytes
  into a fresh slice instead of aliasing the packet buffer
- core.go: AnonymizeAddress copies the IP before zeroing IPv6 bytes,
  instead of corrupting the caller's address in place (session_update
  calls it up to 3x on state.Request.ClientAddress)
- core.go: selection percent boundary now uses >= so SelectionPercent
  selects exactly that percentage of users instead of 1% extra
- relay_packets.go: removed always-false < 0 checks on unsigned counts

+++ modules/common/route_matrix.go
@@ -342,8 +342,8 @@ func (m *RouteMatrix) Analyze() RouteMatrixAnalysis {
 	analysis.AverageRouteLength = float32(averageRouteLength)
 
 	if relayPairs > 0 {
-		analysis.NoRoutePercent = float32(relayPairsWithNoRoutes/relayPairs) * 100.0
-		analysis.OneRoutePercent = float32(relayPairsWithOneRoute/relayPairs) * 100.0
+		analysis.NoRoutePercent = float32(float64(relayPairsWithNoRoutes)/float64(relayPairs)) * 100.0
+		analysis.OneRoutePercent = float32(float64(relayPairsWithOneRoute)/float64(relayPairs)) * 100.0
 	}
 
 	return analysis
